refactor: ValidationErrorResponse, ErrorResponse 클래스 통합
